requirements.txt
Flask==3.0.3
gradio_client==1.4.2
opencv_contrib_python==4.10.0.82
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.74
python-dotenv==1.0.1
Requests==2.32.3
twilio==9.3.4
